migrate to postgreSQL

+++ medvoll/src/main/resources/db/migration/V4__alter-table-medicos-add-column-ativo.sql
@@ -1,2 +1,2 @@
-alter table medicos add ativo tinyint;
+alter table medicos add ativo bigint;
 update medicos set ativo = 1;
